requirements.txt
peewee
pytest
django==4.1.2
djangorestframework